backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==3.2.0",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
